SELECTIONS-PYTHON
PROBLEM STATEMENT:
Finding out the number of possible scenarios in a given situation.

PROJECT DESCRIPTION:
We often come across situations wherein we need to choose a certain scenario,
for example, forming a team - who must be included and who should be not;
corporate conference meetings - where should the project head sit, whether,
in a row or round table etc., it gets tough to calculate the number of
possibilities to do so. Thus, our project would use the concept of permutation
and combination help the user figure out in how many ways could he make
teams, organize birthday parties and functions, decide on seating
arrangement in conferences or professional dinner parties, the number of
ways in which few prizes could be distributed among a certain number of
people etc.